move parser.infra to its own file
